Merge master into develop via merge-master-to-develop

+++ tests/framework/cli/test_pipeline.py
@@ -25,11 +25,13 @@
 #
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
+# pylint: disable=too-many-lines
+import filecmp
 import shutil
 from pathlib import Path
 from zipfile import ZipFile
 
+import click
 import pytest
 import yaml
 from click.testing import CliRunner
@@ -63,7 +65,8 @@ def cleanup_pipelines(dummy_project):
             shutil.rmtree(str(each))
 
         tests = dummy_project / "src" / "tests" / "pipelines" / pipeline
-        shutil.rmtree(str(tests))
+        if tests.is_dir():
+            shutil.rmtree(str(tests))
 
 
 @pytest.fixture
@@ -416,7 +419,7 @@ class TestPipelineDeleteCommand:
     def test_pipeline_delete_confirmation(self, dummy_project, fake_kedro_cli, input_):
         """Test that user confirmation of deletion works"""
         result = CliRunner().invoke(
-            fake_kedro_cli.cli, ["pipeline", "delete", PIPELINE_NAME], input=input_,
+            fake_kedro_cli.cli, ["pipeline", "delete", PIPELINE_NAME], input=input_
         )
 
         source_path = dummy_project / "src" / PACKAGE_NAME / "pipelines" / PIPELINE_NAME
@@ -452,7 +455,7 @@ class TestPipelineDeleteCommand:
         source_path = dummy_project / "src" / PACKAGE_NAME / "pipelines" / PIPELINE_NAME
         shutil.rmtree(str(source_path))
         result = CliRunner().invoke(
-            fake_kedro_cli.cli, ["pipeline", "delete", PIPELINE_NAME], input=input_,
+            fake_kedro_cli.cli, ["pipeline", "delete", PIPELINE_NAME], input=input_
         )
 
         tests_path = dummy_project / "src" / "tests" / "pipelines" / PIPELINE_NAME
@@ -556,7 +559,7 @@ class TestPipelinePackageCommand:
 
     @pytest.mark.parametrize("existing_dir", [True, False])
     def test_pipeline_package_to_destination(
-        self, fake_kedro_cli, existing_dir, tmp_path,
+        self, fake_kedro_cli, existing_dir, tmp_path
     ):
         destination = (tmp_path / "in" / "here").resolve()
         if existing_dir:
@@ -592,9 +595,7 @@ class TestPipelinePackageCommand:
         }
         assert expected_files <= wheel_contents
 
-    def test_pipeline_package_overwrites_wheel(
-        self, fake_kedro_cli, tmp_path,
-    ):
+    def test_pipeline_package_overwrites_wheel(self, fake_kedro_cli, tmp_path):
         destination = (tmp_path / "in" / "here").resolve()
         destination.mkdir(parents=True)
         wheel_file = destination / f"{PIPELINE_NAME}-0.1-py3-none-any.whl"
@@ -675,6 +676,387 @@ class TestPipelinePackageCommand:
         }
         assert expected_files <= wheel_contents
         assert f"{PIPELINE_NAME}/config/parameters.yml" not in wheel_contents
+
+
+@pytest.mark.usefixtures("chdir_to_dummy_project", "patch_log", "cleanup_dist")
+class TestPipelinePullCommand:
+    def call_pipeline_create(self, fake_kedro_cli):
+        result = CliRunner().invoke(
+            fake_kedro_cli.cli, ["pipeline", "create", PIPELINE_NAME]
+        )
+        assert result.exit_code == 0
+
+    def call_pipeline_package(self, fake_kedro_cli, alias=None, destination=None):
+        options = ["--alias", alias] if alias else []
+        options += ["--destination", str(destination)] if destination else []
+        result = CliRunner().invoke(
+            fake_kedro_cli.cli, ["pipeline", "package", PIPELINE_NAME, *options]
+        )
+        assert result.exit_code == 0
+
+    def call_pipeline_delete(self, fake_kedro_cli):
+        result = CliRunner().invoke(
+            fake_kedro_cli.cli, ["pipeline", "delete", "-y", PIPELINE_NAME]
+        )
+        assert result.exit_code == 0
+
+    def assert_package_files_exist(self, source_path):
+        assert {f.name for f in source_path.iterdir()} == {
+            "__init__.py",
+            "nodes.py",
+            "pipeline.py",
+            "README.md",
+        }
+
+    @pytest.mark.parametrize("env", [None, "local"])
+    @pytest.mark.parametrize("alias", [None, "alias_path"])
+    def test_pull_local_whl(self, fake_kedro_cli, dummy_project, env, alias):
+        """
+        Test for pulling a valid wheel file locally.
+        """
+        # pylint: disable=too-many-locals
+        self.call_pipeline_create(fake_kedro_cli)
+        self.call_pipeline_package(fake_kedro_cli)
+        self.call_pipeline_delete(fake_kedro_cli)
+
+        source_path = dummy_project / "src" / PACKAGE_NAME / "pipelines" / PIPELINE_NAME
+        config_path = dummy_project / "conf" / "base" / "pipelines" / PIPELINE_NAME
+        test_path = dummy_project / "src" / "tests" / "pipelines" / PIPELINE_NAME
+        # Make sure the files actually deleted before pulling from the wheel file.
+        assert not source_path.exists()
+        assert not test_path.exists()
+        assert not config_path.exists()
+
+        wheel_file = (
+            dummy_project / "src" / "dist" / f"{PIPELINE_NAME}-0.1-py3-none-any.whl"
+        )
+        assert wheel_file.is_file()
+
+        options = ["-e", env] if env else []
+        options += ["--alias", alias] if alias else []
+        result = CliRunner().invoke(
+            fake_kedro_cli.cli, ["pipeline", "pull", str(wheel_file), *options]
+        )
+        assert result.exit_code == 0
+
+        pipeline_name = alias or PIPELINE_NAME
+        source_dest = dummy_project / "src" / PACKAGE_NAME / "pipelines" / pipeline_name
+        config_env = env or "base"
+        config_dest = dummy_project / "conf" / config_env / "pipelines" / pipeline_name
+        test_dest = dummy_project / "src" / "tests" / "pipelines" / pipeline_name
+
+        self.assert_package_files_exist(source_dest)
+        assert (config_dest / "parameters.yml").is_file()
+        assert {f.name for f in test_dest.iterdir()} == {
+            "__init__.py",
+            "test_pipeline.py",
+        }
+
+    @pytest.mark.parametrize("env", [None, "local"])
+    @pytest.mark.parametrize("alias", [None, "alias_path"])
+    def test_pull_local_whl_compare(self, fake_kedro_cli, dummy_project, env, alias):
+        """
+        Test for pulling a valid wheel file locally, unpack it into another location and
+        check that unpacked files are identical to the ones in the original modular pipeline.
+        """
+        # pylint: disable=too-many-locals
+        pipeline_name = "another_pipeline"
+        self.call_pipeline_create(fake_kedro_cli)
+        self.call_pipeline_package(fake_kedro_cli, pipeline_name)
+
+        source_path = dummy_project / "src" / PACKAGE_NAME / "pipelines" / PIPELINE_NAME
+        config_path = dummy_project / "conf" / "base" / "pipelines" / PIPELINE_NAME
+        test_path = dummy_project / "src" / "tests" / "pipelines" / PIPELINE_NAME
+
+        wheel_file = (
+            dummy_project / "src" / "dist" / f"{pipeline_name}-0.1-py3-none-any.whl"
+        )
+        assert wheel_file.is_file()
+
+        options = ["-e", env] if env else []
+        options += ["--alias", alias] if alias else []
+        result = CliRunner().invoke(
+            fake_kedro_cli.cli, ["pipeline", "pull", str(wheel_file), *options]
+        )
+        assert result.exit_code == 0
+
+        pipeline_name = alias or pipeline_name
+        source_dest = dummy_project / "src" / PACKAGE_NAME / "pipelines" / pipeline_name
+        config_env = env or "base"
+        config_dest = dummy_project / "conf" / config_env / "pipelines" / pipeline_name
+        test_dest = dummy_project / "src" / "tests" / "pipelines" / pipeline_name
+
+        assert not filecmp.dircmp(source_path, source_dest).diff_files
+        assert not filecmp.dircmp(config_path, config_dest).diff_files
+        assert not filecmp.dircmp(test_path, test_dest).diff_files
+
+    def test_pull_two_dist_info(self, fake_kedro_cli, dummy_project, mocker, tmp_path):
+        """
+        Test for pulling a wheel file with more than one dist-info directory.
+        """
+        self.call_pipeline_package(fake_kedro_cli)
+        wheel_file = (
+            dummy_project / "src" / "dist" / f"{PIPELINE_NAME}-0.1-py3-none-any.whl"
+        )
+        assert wheel_file.is_file()
+
+        (tmp_path / "dummy.dist-info").mkdir()
+
+        mocker.patch(
+            "kedro.framework.cli.pipeline.tempfile.TemporaryDirectory",
+            return_value=tmp_path,
+        )
+        result = CliRunner().invoke(
+            fake_kedro_cli.cli, ["pipeline", "pull", str(wheel_file)]
+        )
+        assert result.exit_code
+        assert "Error: More than 1 or no dist-info files found" in result.output
+
+    @pytest.mark.parametrize("env", [None, "local"])
+    @pytest.mark.parametrize("alias", [None, "alias_path"])
+    def test_pull_tests_missing(self, fake_kedro_cli, dummy_project, env, alias):
+        """
+        Test for pulling a valid wheel file locally, but `tests` directory is missing
+        from the wheel file.
+        """
+        # pylint: disable=too-many-locals
+        self.call_pipeline_create(fake_kedro_cli)
+        test_path = dummy_project / "src" / "tests" / "pipelines" / PIPELINE_NAME
+        shutil.rmtree(test_path)
+        assert not test_path.exists()
+        self.call_pipeline_package(fake_kedro_cli)
+        self.call_pipeline_delete(fake_kedro_cli)
+
+        source_path = dummy_project / "src" / PACKAGE_NAME / "pipelines" / PIPELINE_NAME
+        config_path = dummy_project / "conf" / "base" / "pipelines" / PIPELINE_NAME
+        # Make sure the files actually deleted before pulling from the wheel file.
+        assert not source_path.exists()
+        assert not config_path.exists()
+
+        wheel_file = (
+            dummy_project / "src" / "dist" / f"{PIPELINE_NAME}-0.1-py3-none-any.whl"
+        )
+        assert wheel_file.is_file()
+
+        options = ["-e", env] if env else []
+        options += ["--alias", alias] if alias else []
+        result = CliRunner().invoke(
+            fake_kedro_cli.cli, ["pipeline", "pull", str(wheel_file), *options]
+        )
+        assert result.exit_code == 0
+
+        pipeline_name = alias or PIPELINE_NAME
+        source_dest = dummy_project / "src" / PACKAGE_NAME / "pipelines" / pipeline_name
+        config_env = env or "base"
+        config_dest = dummy_project / "conf" / config_env / "pipelines" / pipeline_name
+        test_dest = dummy_project / "src" / "tests" / "pipelines" / pipeline_name
+
+        self.assert_package_files_exist(source_dest)
+        assert (config_dest / "parameters.yml").is_file()
+        assert not test_dest.exists()
+
+    @pytest.mark.parametrize("env", [None, "local"])
+    @pytest.mark.parametrize("alias", [None, "alias_path"])
+    def test_pull_config_missing(self, fake_kedro_cli, dummy_project, env, alias):
+        """
+        Test for pulling a valid wheel file locally, but `config` directory is missing
+        from the wheel file.
+        """
+        # pylint: disable=too-many-locals
+        self.call_pipeline_create(fake_kedro_cli)
+        config_path = dummy_project / "conf" / "base" / "pipelines" / PIPELINE_NAME
+        shutil.rmtree(config_path)
+        assert not config_path.exists()
+        self.call_pipeline_package(fake_kedro_cli)
+        self.call_pipeline_delete(fake_kedro_cli)
+
+        source_path = dummy_project / "src" / PACKAGE_NAME / "pipelines" / PIPELINE_NAME
+        test_path = dummy_project / "src" / "tests" / "pipelines" / PIPELINE_NAME
+        # Make sure the files actually deleted before pulling from the wheel file.
+        assert not source_path.exists()
+        assert not test_path.exists()
+
+        wheel_file = (
+            dummy_project / "src" / "dist" / f"{PIPELINE_NAME}-0.1-py3-none-any.whl"
+        )
+        assert wheel_file.is_file()
+
+        options = ["-e", env] if env else []
+        options += ["--alias", alias] if alias else []
+        result = CliRunner().invoke(
+            fake_kedro_cli.cli, ["pipeline", "pull", str(wheel_file), *options]
+        )
+        assert result.exit_code == 0
+
+        pipeline_name = alias or PIPELINE_NAME
+        source_dest = dummy_project / "src" / PACKAGE_NAME / "pipelines" / pipeline_name
+        config_env = env or "base"
+        config_dest = dummy_project / "conf" / config_env / "pipelines" / pipeline_name
+        test_dest = dummy_project / "src" / "tests" / "pipelines" / pipeline_name
+
+        self.assert_package_files_exist(source_dest)
+        assert not config_dest.exists()
+        assert {f.name for f in test_dest.iterdir()} == {
+            "__init__.py",
+            "test_pipeline.py",
+        }
+
+    @pytest.mark.parametrize("env", [None, "local"])
+    @pytest.mark.parametrize("alias", [None, "alias_path"])
+    def test_pull_pipeline_missing(self, fake_kedro_cli, dummy_project, env, alias):
+        """
+        Test for pulling a valid wheel file locally, but `pipeline.py` is missing from the wheel
+        file.
+        """
+        # pylint: disable=too-many-locals
+        self.call_pipeline_create(fake_kedro_cli)
+        package_path = (
+            dummy_project / "src" / PACKAGE_NAME / "pipelines" / PIPELINE_NAME
+        )
+        shutil.rmtree(package_path)
+        assert not package_path.exists()
+        self.call_pipeline_package(fake_kedro_cli)
+        self.call_pipeline_delete(fake_kedro_cli)
+
+        config_path = dummy_project / "conf" / "base" / "pipelines" / PIPELINE_NAME
+        test_path = dummy_project / "src" / "tests" / "pipelines" / PIPELINE_NAME
+        # Make sure the files actually deleted before pulling from the wheel file.
+        assert not test_path.exists()
+        assert not config_path.exists()
+
+        wheel_file = (
+            dummy_project / "src" / "dist" / f"{PIPELINE_NAME}-0.1-py3-none-any.whl"
+        )
+        assert wheel_file.is_file()
+
+        options = ["-e", env] if env else []
+        options += ["--alias", alias] if alias else []
+        result = CliRunner().invoke(
+            fake_kedro_cli.cli, ["pipeline", "pull", str(wheel_file), *options]
+        )
+        assert not result.exit_code
+
+        pipeline_name = alias or PIPELINE_NAME
+        source_dest = dummy_project / "src" / PACKAGE_NAME / "pipelines" / pipeline_name
+        config_env = env or "base"
+        config_dest = dummy_project / "conf" / config_env / "pipelines" / pipeline_name
+        test_dest = dummy_project / "src" / "tests" / "pipelines" / pipeline_name
+
+        assert not source_dest.exists()
+        assert not (config_dest / "parameters.yml").is_file()
+        assert {f.name for f in test_dest.iterdir()} == {
+            "__init__.py",
+            "test_pipeline.py",
+        }
+
+    @pytest.mark.parametrize("env", [None, "local"])
+    @pytest.mark.parametrize("alias", [None, "alias_path"])
+    def test_pull_from_pypi(
+        self, fake_kedro_cli, dummy_project, mocker, tmp_path, env, alias
+    ):
+        """
+        Test for pulling a valid wheel file from pypi.
+        """
+        # pylint: disable=too-many-locals
+        self.call_pipeline_create(fake_kedro_cli)
+        # We mock the `pip download` call, and manually create a package wheel file
+        # to simulate the pypi scenario instead
+        self.call_pipeline_package(fake_kedro_cli, destination=tmp_path)
+        wheel_file = tmp_path / f"{PIPELINE_NAME}-0.1-py3-none-any.whl"
+        assert wheel_file.is_file()
+        self.call_pipeline_delete(fake_kedro_cli)
+
+        source_path = dummy_project / "src" / PACKAGE_NAME / "pipelines" / PIPELINE_NAME
+        config_path = dummy_project / "conf" / "base" / "pipelines" / PIPELINE_NAME
+        test_path = dummy_project / "src" / "tests" / "pipelines" / PIPELINE_NAME
+        # Make sure the files actually deleted before pulling from pypi.
+        assert not source_path.exists()
+        assert not test_path.exists()
+        assert not config_path.exists()
+
+        python_call_mock = mocker.patch("kedro.framework.cli.pipeline.python_call")
+        mocker.patch(
+            "kedro.framework.cli.pipeline.tempfile.TemporaryDirectory",
+            return_value=tmp_path,
+        )
+
+        options = ["-e", env] if env else []
+        options += ["--alias", alias] if alias else []
+        result = CliRunner().invoke(
+            fake_kedro_cli.cli, ["pipeline", "pull", PIPELINE_NAME, *options]
+        )
+        assert result.exit_code == 0
+
+        python_call_mock.assert_called_once_with(
+            "pip", ["download", "--no-deps", "--dest", str(tmp_path), PIPELINE_NAME]
+        )
+
+        pipeline_name = alias or PIPELINE_NAME
+        source_dest = dummy_project / "src" / PACKAGE_NAME / "pipelines" / pipeline_name
+        config_env = env or "base"
+        config_dest = dummy_project / "conf" / config_env / "pipelines" / pipeline_name
+        test_dest = dummy_project / "src" / "tests" / "pipelines" / pipeline_name
+
+        self.assert_package_files_exist(source_dest)
+        assert (config_dest / "parameters.yml").is_file()
+        assert {f.name for f in test_dest.iterdir()} == {
+            "__init__.py",
+            "test_pipeline.py",
+        }
+
+    def test_invalid_pull_from_pypi(self, fake_kedro_cli, mocker, tmp_path):
+        """
+        Test for pulling package from pypi, and it cannot be found.
+        """
+
+        pypi_error_message = (
+            "ERROR: Could not find a version that satisfies the requirement"
+        )
+        python_call_mock = mocker.patch(
+            "kedro.framework.cli.pipeline.python_call",
+            side_effect=click.ClickException(pypi_error_message),
+        )
+        mocker.patch(
+            "kedro.framework.cli.pipeline.tempfile.TemporaryDirectory",
+            return_value=tmp_path,
+        )
+
+        invalid_pypi_name = "non_existent"
+        result = CliRunner().invoke(
+            fake_kedro_cli.cli, ["pipeline", "pull", invalid_pypi_name]
+        )
+        assert result.exit_code
+
+        python_call_mock.assert_called_once_with(
+            "pip", ["download", "--no-deps", "--dest", str(tmp_path), invalid_pypi_name]
+        )
+
+        assert pypi_error_message in result.stdout
+
+    def test_pull_from_pypi_more_than_one_wheel_file(
+        self, fake_kedro_cli, mocker, tmp_path
+    ):
+        """
+        Test for pulling a wheel file with `pip download`, but there are more than one wheel
+        file to unzip.
+        """
+        # We mock the `pip download` call, and manually create a package wheel file
+        # to simulate the pypi scenario instead
+        self.call_pipeline_package(fake_kedro_cli, destination=tmp_path)
+        self.call_pipeline_package(
+            fake_kedro_cli, alias="another", destination=tmp_path
+        )
+        mocker.patch("kedro.framework.cli.pipeline.python_call")
+        mocker.patch(
+            "kedro.framework.cli.pipeline.tempfile.TemporaryDirectory",
+            return_value=tmp_path,
+        )
+        result = CliRunner().invoke(
+            fake_kedro_cli.cli, ["pipeline", "pull", PIPELINE_NAME]
+        )
+
+        assert result.exit_code
+        assert "Error: More than 1 or no wheel files found:" in result.output
 
 
 class TestSyncDirs:
